requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0

Babel>=1.3
cliff>=1.10.0  # Apache-2.0
iso8601>=0.1.9
PrettyTable>=0.7,<0.8
python-keystoneclient>=1.1.0
python-openstackclient>=1.0.0
requests>=2.2.0,!=2.4.0
simplejson>=2.2.0
six>=1.9.0
